feat: PostUseCase 클래스 구현 테스트 작성 passed 30 of 30 트랜잭션 처리 (@transactional) 도메인 이벤트 발행 예외 처리 및 검증
 커밋할 변경 사항:
	새 파일:       src/main/java/backend/airo/application/post/usecase/PostUseCase.java
	수정함:        src/main/java/backend/airo/domain/post/Post.java
	수정함:        src/main/java/backend/airo/domain/post/command/CreatePostCommand.java
	수정함:        src/main/java/backend/airo/domain/post/command/UpdatePostCommand.java
	새 파일:       src/main/java/backend/airo/domain/post/exception/PostAccessDeniedException.java
	새 파일:       src/main/java/backend/airo/domain/post/exception/PostDeleteException.java

+++ src/main/java/backend/airo/domain/post/command/CreatePostCommand.java
@@ -24,7 +24,7 @@ public record CreatePostCommand(
         PostStatus status,
 
         @Positive(message = "카테고리 ID는 양수여야 합니다")
-        Integer categoryId,
+        Long categoryId,
 
         @Positive(message = "위치 ID는 양수여야 합니다")
         Long locationId,
@@ -71,7 +71,7 @@ public record CreatePostCommand(
      * 발행 커맨드 생성
      */
     public static CreatePostCommand forPublish(String title, String content, Long userId,
-                                               Integer categoryId, Long locationId) {
+                                               Long categoryId, Long locationId) {
         return new CreatePostCommand(
                 title, content, userId, PostStatus.PUBLISHED,
                 categoryId, locationId, null, List.of(), List.of(), false
